backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
requests>=2.31.0
opencv-python>=4.8.0
ultralytics>=8.0.0
psutil>=5.9.0
websockets>=11.0.3
colorama>=0.4.6
